requirements.txt
fastapi
uvicorn
requests
cachetools
slowapi
python-dotenv
beautifulsoup4
lxml
trafilatura
